Refactor CommunityOnchainTransactions by consolidating transaction type imports and removing unused SCSS file for improved clarity and maintainability.
