--- packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityOnchainTransactions/CommunityOnchainTransactions.scss
@@ -1,8 +1,0 @@
-.CommunityTransactions {
-  width: 100%;
-  max-width: 1104px;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
